refactor(gateway): extract non-agent reply finalization
